C functions to be used by other packages.
To make the C functions available to call from your
packages’ C code, you have to:
- Add cubature to LinkingTo in your packages’ DESCRIPTION file.
- Add #include <cubature.h> in your C or C++ code.

If you get a duplicated symbol error in while linking,
consider adding #include <cubature-typedefs.h> in 
your h file, and the other include in your c, cpp
file only.

The namespace is loaded automatically the first time
a C function in cubature.h is called, so that there 
is no need to add cubature to your packages’ Depends
and adding dummy imports (in case they are not there 
for other reasons).